File name changes, and moved solver to its own file.
